Isolate `pr status` command
