+++ command/pr_test.go
@@ -21,276 +21,6 @@ func eq(t *testing.T, got interface{}, expected interface{}) {
 	}
 }
 
-func TestPRStatus(t *testing.T) {
-	initBlankContext("", "OWNER/REPO", "blueberries")
-	http := initFakeHTTP()
-	http.StubRepoResponse("OWNER", "REPO")
-	http.Register(httpmock.GraphQL(`query PullRequestStatus\b`), httpmock.FileResponse("../test/fixtures/prStatus.json"))
-
-	output, err := RunCommand("pr status")
-	if err != nil {
-		t.Errorf("error running command `pr status`: %v", err)
-	}
-
-	expectedPrs := []*regexp.Regexp{
-		regexp.MustCompile(`#8.*\[strawberries\]`),
-		regexp.MustCompile(`#9.*\[apples\]`),
-		regexp.MustCompile(`#10.*\[blueberries\]`),
-		regexp.MustCompile(`#11.*\[figs\]`),
-	}
-
-	for _, r := range expectedPrs {
-		if !r.MatchString(output.String()) {
-			t.Errorf("output did not match regexp /%s/", r)
-		}
-	}
-}
-
-func TestPRStatus_fork(t *testing.T) {
-	initBlankContext("", "OWNER/REPO", "blueberries")
-	http := initFakeHTTP()
-	http.StubForkedRepoResponse("OWNER/REPO", "PARENT/REPO")
-	http.Register(httpmock.GraphQL(`query PullRequestStatus\b`), httpmock.FileResponse("../test/fixtures/prStatusFork.json"))
-
-	defer run.SetPrepareCmd(func(cmd *exec.Cmd) run.Runnable {
-		switch strings.Join(cmd.Args, " ") {
-		case `git config --get-regexp ^branch\.blueberries\.(remote|merge)$`:
-			return &test.OutputStub{Out: []byte(`branch.blueberries.remote origin
-branch.blueberries.merge refs/heads/blueberries`)}
-		default:
-			panic("not implemented")
-		}
-	})()
-
-	output, err := RunCommand("pr status")
-	if err != nil {
-		t.Fatalf("error running command `pr status`: %v", err)
-	}
-
-	branchRE := regexp.MustCompile(`#10.*\[OWNER:blueberries\]`)
-	if !branchRE.MatchString(output.String()) {
-		t.Errorf("did not match current branch:\n%v", output.String())
-	}
-}
-
-func TestPRStatus_reviewsAndChecks(t *testing.T) {
-	initBlankContext("", "OWNER/REPO", "blueberries")
-	http := initFakeHTTP()
-	http.StubRepoResponse("OWNER", "REPO")
-	http.Register(httpmock.GraphQL(`query PullRequestStatus\b`), httpmock.FileResponse("../test/fixtures/prStatusChecks.json"))
-
-	output, err := RunCommand("pr status")
-	if err != nil {
-		t.Errorf("error running command `pr status`: %v", err)
-	}
-
-	expected := []string{
-		"✓ Checks passing + Changes requested",
-		"- Checks pending ✓ Approved",
-		"× 1/3 checks failing - Review required",
-	}
-
-	for _, line := range expected {
-		if !strings.Contains(output.String(), line) {
-			t.Errorf("output did not contain %q: %q", line, output.String())
-		}
-	}
-}
-
-func TestPRStatus_currentBranch_showTheMostRecentPR(t *testing.T) {
-	initBlankContext("", "OWNER/REPO", "blueberries")
-	http := initFakeHTTP()
-	http.StubRepoResponse("OWNER", "REPO")
-	http.Register(httpmock.GraphQL(`query PullRequestStatus\b`), httpmock.FileResponse("../test/fixtures/prStatusCurrentBranch.json"))
-
-	output, err := RunCommand("pr status")
-	if err != nil {
-		t.Errorf("error running command `pr status`: %v", err)
-	}
-
-	expectedLine := regexp.MustCompile(`#10  Blueberries are certainly a good fruit \[blueberries\]`)
-	if !expectedLine.MatchString(output.String()) {
-		t.Errorf("output did not match regexp /%s/\n> output\n%s\n", expectedLine, output)
-		return
-	}
-
-	unexpectedLines := []*regexp.Regexp{
-		regexp.MustCompile(`#9  Blueberries are a good fruit \[blueberries\] - Merged`),
-		regexp.MustCompile(`#8  Blueberries are probably a good fruit \[blueberries\] - Closed`),
-	}
-	for _, r := range unexpectedLines {
-		if r.MatchString(output.String()) {
-			t.Errorf("output unexpectedly match regexp /%s/\n> output\n%s\n", r, output)
-			return
-		}
-	}
-}
-
-func TestPRStatus_currentBranch_defaultBranch(t *testing.T) {
-	initBlankContext("", "OWNER/REPO", "blueberries")
-	http := initFakeHTTP()
-	http.StubRepoResponseWithDefaultBranch("OWNER", "REPO", "blueberries")
-	http.Register(httpmock.GraphQL(`query PullRequestStatus\b`), httpmock.FileResponse("../test/fixtures/prStatusCurrentBranch.json"))
-
-	output, err := RunCommand("pr status")
-	if err != nil {
-		t.Errorf("error running command `pr status`: %v", err)
-	}
-
-	expectedLine := regexp.MustCompile(`#10  Blueberries are certainly a good fruit \[blueberries\]`)
-	if !expectedLine.MatchString(output.String()) {
-		t.Errorf("output did not match regexp /%s/\n> output\n%s\n", expectedLine, output)
-		return
-	}
-}
-
-func TestPRStatus_currentBranch_defaultBranch_repoFlag(t *testing.T) {
-	initBlankContext("", "OWNER/REPO", "blueberries")
-	http := initFakeHTTP()
-	http.Register(httpmock.GraphQL(`query PullRequestStatus\b`), httpmock.FileResponse("../test/fixtures/prStatusCurrentBranchClosedOnDefaultBranch.json"))
-
-	output, err := RunCommand("pr status -R OWNER/REPO")
-	if err != nil {
-		t.Errorf("error running command `pr status`: %v", err)
-	}
-
-	expectedLine := regexp.MustCompile(`#8  Blueberries are a good fruit \[blueberries\]`)
-	if expectedLine.MatchString(output.String()) {
-		t.Errorf("output not expected to match regexp /%s/\n> output\n%s\n", expectedLine, output)
-		return
-	}
-}
-
-func TestPRStatus_currentBranch_Closed(t *testing.T) {
-	initBlankContext("", "OWNER/REPO", "blueberries")
-	http := initFakeHTTP()
-	http.StubRepoResponse("OWNER", "REPO")
-	http.Register(httpmock.GraphQL(`query PullRequestStatus\b`), httpmock.FileResponse("../test/fixtures/prStatusCurrentBranchClosed.json"))
-
-	output, err := RunCommand("pr status")
-	if err != nil {
-		t.Errorf("error running command `pr status`: %v", err)
-	}
-
-	expectedLine := regexp.MustCompile(`#8  Blueberries are a good fruit \[blueberries\] - Closed`)
-	if !expectedLine.MatchString(output.String()) {
-		t.Errorf("output did not match regexp /%s/\n> output\n%s\n", expectedLine, output)
-		return
-	}
-}
-
-func TestPRStatus_currentBranch_Closed_defaultBranch(t *testing.T) {
-	initBlankContext("", "OWNER/REPO", "blueberries")
-	http := initFakeHTTP()
-	http.StubRepoResponseWithDefaultBranch("OWNER", "REPO", "blueberries")
-	http.Register(httpmock.GraphQL(`query PullRequestStatus\b`), httpmock.FileResponse("../test/fixtures/prStatusCurrentBranchClosedOnDefaultBranch.json"))
-
-	output, err := RunCommand("pr status")
-	if err != nil {
-		t.Errorf("error running command `pr status`: %v", err)
-	}
-
-	expectedLine := regexp.MustCompile(`There is no pull request associated with \[blueberries\]`)
-	if !expectedLine.MatchString(output.String()) {
-		t.Errorf("output did not match regexp /%s/\n> output\n%s\n", expectedLine, output)
-		return
-	}
-}
-
-func TestPRStatus_currentBranch_Merged(t *testing.T) {
-	initBlankContext("", "OWNER/REPO", "blueberries")
-	http := initFakeHTTP()
-	http.StubRepoResponse("OWNER", "REPO")
-	http.Register(httpmock.GraphQL(`query PullRequestStatus\b`), httpmock.FileResponse("../test/fixtures/prStatusCurrentBranchMerged.json"))
-
-	output, err := RunCommand("pr status")
-	if err != nil {
-		t.Errorf("error running command `pr status`: %v", err)
-	}
-
-	expectedLine := regexp.MustCompile(`#8  Blueberries are a good fruit \[blueberries\] - Merged`)
-	if !expectedLine.MatchString(output.String()) {
-		t.Errorf("output did not match regexp /%s/\n> output\n%s\n", expectedLine, output)
-		return
-	}
-}
-
-func TestPRStatus_currentBranch_Merged_defaultBranch(t *testing.T) {
-	initBlankContext("", "OWNER/REPO", "blueberries")
-	http := initFakeHTTP()
-	http.StubRepoResponseWithDefaultBranch("OWNER", "REPO", "blueberries")
-	http.Register(httpmock.GraphQL(`query PullRequestStatus\b`), httpmock.FileResponse("../test/fixtures/prStatusCurrentBranchMergedOnDefaultBranch.json"))
-
-	output, err := RunCommand("pr status")
-	if err != nil {
-		t.Errorf("error running command `pr status`: %v", err)
-	}
-
-	expectedLine := regexp.MustCompile(`There is no pull request associated with \[blueberries\]`)
-	if !expectedLine.MatchString(output.String()) {
-		t.Errorf("output did not match regexp /%s/\n> output\n%s\n", expectedLine, output)
-		return
-	}
-}
-
-func TestPRStatus_blankSlate(t *testing.T) {
-	initBlankContext("", "OWNER/REPO", "blueberries")
-	http := initFakeHTTP()
-	http.StubRepoResponse("OWNER", "REPO")
-	http.Register(httpmock.GraphQL(`query PullRequestStatus\b`), httpmock.StringResponse(`{"data": {}}`))
-
-	output, err := RunCommand("pr status")
-	if err != nil {
-		t.Errorf("error running command `pr status`: %v", err)
-	}
-
-	expected := `
-Relevant pull requests in OWNER/REPO
-
-Current branch
-  There is no pull request associated with [blueberries]
-
-Created by you
-  You have no open pull requests
-
-Requesting a code review from you
-  You have no pull requests to review
-
-`
-	if output.String() != expected {
-		t.Errorf("expected %q, got %q", expected, output.String())
-	}
-}
-
-func TestPRStatus_detachedHead(t *testing.T) {
-	initBlankContext("", "OWNER/REPO", "")
-	http := initFakeHTTP()
-	http.StubRepoResponse("OWNER", "REPO")
-	http.Register(httpmock.GraphQL(`query PullRequestStatus\b`), httpmock.StringResponse(`{"data": {}}`))
-
-	output, err := RunCommand("pr status")
-	if err != nil {
-		t.Errorf("error running command `pr status`: %v", err)
-	}
-
-	expected := `
-Relevant pull requests in OWNER/REPO
-
-Current branch
-  There is no current branch
-
-Created by you
-  You have no open pull requests
-
-Requesting a code review from you
-  You have no pull requests to review
-
-`
-	if output.String() != expected {
-		t.Errorf("expected %q, got %q", expected, output.String())
-	}
-}
-
 func TestPRList(t *testing.T) {
 	initBlankContext("", "OWNER/REPO", "master")
 	defer stubTerminal(true)()
@@ -479,13 +209,6 @@ func TestPRList_web(t *testing.T) {
 	}
 	url := seenCmd.Args[len(seenCmd.Args)-1]
 	eq(t, url, expectedURL)
-}
-
-func TestReplaceExcessiveWhitespace(t *testing.T) {
-	eq(t, replaceExcessiveWhitespace("hello\ngoodbye"), "hello goodbye")
-	eq(t, replaceExcessiveWhitespace("  hello goodbye  "), "hello goodbye")
-	eq(t, replaceExcessiveWhitespace("hello   goodbye"), "hello goodbye")
-	eq(t, replaceExcessiveWhitespace("   hello \n goodbye   "), "hello goodbye")
 }
 
 func TestPrClose(t *testing.T) {
