+++ pkg/cmd/pr/status/fixtures/prStatusCurrentBranchClosed.json
