+++ pkg/cmd/pr/status/fixtures/prStatusCurrentBranchClosedOnDefaultBranch.json
