+++ pkg/cmd/pr/status/fixtures/prStatusCurrentBranchMerged.json
