+++ pkg/cmd/pr/status/fixtures/prStatusCurrentBranchMergedOnDefaultBranch.json
