+++ pkg/text/sanitize.go
@@ -1,0 +1,12 @@
+package text
+
+import (
+	"regexp"
+	"strings"
+)
+
+var ws = regexp.MustCompile(`\s+`)
+
+func ReplaceExcessiveWhitespace(s string) string {
+	return ws.ReplaceAllString(strings.TrimSpace(s), " ")
+}
